# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"group8-bootcamp-finalproject","description":"bootcamp-final-ismael-piloto","contact":{"name":"bootcamp-final-ismael-piloto","email":"ismael.piloto@mercadolivre.com"},"version":"last"},"servers":[{"url":"https://null_group8-bootcamp-finalproject.furyapps.io","description":"Scope null"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/fresh-products/orders/{idOrder}":{"get":{"tags":["purchase-order-controller"],"operationId":"getOrders","parameters":[{"name":"idOrder","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductListDTO"}}}}}},"put":{"tags":["purchase-order-controller"],"operationId":"editOrder","parameters":[{"name":"idOrder","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderPriceResponseDTO"}}}}}}},"/api/v1/fresh-products/orders/cancel/{idOrder}":{"put":{"tags":["purchase-order-controller"],"operationId":"cancelOrder","parameters":[{"name":"idOrder","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"204":{"description":"No Content"}}}},"/api/v1/fresh-products/inboundorder":{"put":{"tags":["inbound-order-controller"],"operationId":"updateInboundOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InboundOrderResponseDTO"}}}}}},"post":{"tags":["inbound-order-controller"],"operationId":"createInboundOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InboundOrderResponseDTO"}}}}}}},"/api/v1/sign-in":{"post":{"tags":["session-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponseDTO"}}}}}}},"/api/v1/fresh-products/orders":{"post":{"tags":["purchase-order-controller"],"operationId":"newPurchaseOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderPriceResponseDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products/warehouse/{productId}":{"get":{"tags":["warehouse-controller"],"operationId":"findAllProductsFromWarehouseById","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WarehouseProductListDTO"}}}}}}},"/api/v1/fresh-products/list/{productCategory}":{"get":{"tags":["product-controller"],"operationId":"listProductsByCategory","parameters":[{"name":"productCategory","in":"path","required":true,"schema":{"type":"string","enum":["FS","RF","FF"]}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductListDTO"}}}}}}},"/api/v1/fresh-products/due-date/list/{daysQuantity}":{"get":{"tags":["due-date-controller"],"operationId":"listBatchPerProductCategoryOrderedByDueDate","parameters":[{"name":"daysQuantity","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"productCategory","in":"query","required":false,"schema":{"type":"string","enum":["FS","RF","FF"]}},{"name":"order","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"},"default":"dueDate_desc"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatchStockDueDateListDTO"}}}}}}},"/api/v1/fresh-products/batch/list/{productId}":{"get":{"tags":["batch-controller"],"operationId":"listProductPerBatch","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"order","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"},"default":"dueDate_desc"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductBatchDTO"}}}}}}},"/api/v1/fresh-products/":{"get":{"tags":["product-controller"],"operationId":"listAllProducts","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductListDTO"}}}}}}}},"components":{"schemas":{"ApiError":{"type":"object","properties":{"error":{"type":"string"},"message":{"type":"string"},"status":{"type":"integer","format":"int32"}}},"ValidationError":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}}},"OrderStatusDTO":{"required":["statusCode"],"type":"object","properties":{"statusCode":{"type":"string","enum":["OPEN","CANCELED"]}}},"ProductQuantityRequestDTO":{"required":["productId","quantity"],"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"quantity":{"minimum":1,"type":"integer","format":"int32"}}},"PurchaseOrderDTO":{"required":["buyerId","orderStatus","products"],"type":"object","properties":{"buyerId":{"type":"integer","format":"int64"},"orderStatus":{"$ref":"#/components/schemas/OrderStatusDTO"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductQuantityRequestDTO"}}}},"PurchaseOrderRequestDTO":{"type":"object","properties":{"purchaseOrder":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}},"PurchaseOrderPriceResponseDTO":{"type":"object","properties":{"totalPrice":{"type":"number","format":"double"}}},"BatchDTO":{"required":["batchNumber","currentTemperature","productId","quantity"],"type":"object","properties":{"batchId":{"type":"integer","format":"int64"},"batchNumber":{"type":"string"},"productId":{"type":"integer","format":"int64"},"currentTemperature":{"type":"number","format":"double"},"quantity":{"minimum":1,"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"$ref":"#/components/schemas/LocalTime"},"dueDate":{"type":"string","format":"date"}}},"InboundOrderDTO":{"type":"object","properties":{"inboundOrderId":{"type":"integer","format":"int64"},"section":{"$ref":"#/components/schemas/WarehouseSectionDTO"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchDTO"}}}},"InboundOrderRequestDTO":{"required":["inboundOrder"],"type":"object","properties":{"inboundOrder":{"$ref":"#/components/schemas/InboundOrderDTO"}}},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"WarehouseSectionDTO":{"required":["sectionCode","warehouseCode"],"type":"object","properties":{"sectionCode":{"type":"integer","format":"int64"},"warehouseCode":{"type":"integer","format":"int64"}}},"BatchResponseDTO":{"type":"object","properties":{"batchId":{"type":"integer","format":"int64"},"batchNumber":{"type":"string"},"productId":{"type":"integer","format":"int64"},"currentTemperature":{"type":"number","format":"double"},"minimumTemperature":{"type":"number","format":"double"},"quantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"$ref":"#/components/schemas/LocalTime"},"dueDate":{"type":"string","format":"date"}}},"InboundOrderResponseDTO":{"type":"object","properties":{"inboundOrderId":{"type":"integer","format":"int64"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchResponseDTO"}}}},"LoginRequestDTO":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AccountResponseDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"token":{"type":"string"}}},"WarehouseProductListDTO":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"warehouses":{"type":"array","items":{"$ref":"#/components/schemas/WarehouseTotalProductDTO"}}}},"WarehouseTotalProductDTO":{"type":"object","properties":{"warehouseCode":{"type":"integer","format":"int64"},"totalQuantity":{"type":"integer","format":"int64"}}},"ProductDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"minTemperature":{"type":"number","format":"double"},"maxTemperature":{"type":"number","format":"double"},"price":{"type":"number","format":"double"}}},"ProductListDTO":{"type":"object","properties":{"products":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}},"BatchStockDueDateDTO":{"type":"object","properties":{"sectionId":{"type":"integer","format":"int64"},"batchNumber":{"type":"string"},"productId":{"type":"integer","format":"int64"},"productCategory":{"type":"string"},"dueDate":{"type":"string","format":"date"},"quantity":{"type":"integer","format":"int32"}}},"BatchStockDueDateListDTO":{"type":"object","properties":{"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchStockDueDateDTO"}}}},"BatchStockDTO":{"type":"object","properties":{"wareHouseId":{"type":"integer","format":"int64"},"sectionId":{"type":"integer","format":"int64"},"batchId":{"type":"integer","format":"int64"},"batchNumber":{"type":"string"},"currentQuantity":{"type":"integer","format":"int32"},"dueDate":{"type":"string","format":"date"}}},"ProductBatchDTO":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchStockDTO"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}